+++ spec/settings_reader/vault_resolver/configuration_spec.rb
@@ -1,0 +1,48 @@
+RSpec.describe SettingsReader::VaultResolver::Configuration do
+  let(:config) { described_class.new }
+
+  describe '#setup_lease_refresher' do
+    let(:task) { instance_double(Concurrent::TimerTask, execute: true) }
+
+    before { allow(Concurrent::TimerTask).to receive(:new).and_return(task) }
+
+    context 'without previous task' do
+      it 'instantiate task with correct params' do
+        config.setup_lease_refresher
+        expect(Concurrent::TimerTask).to have_received(:new).with(execution_interval: config.lease_refresh_interval)
+      end
+
+      it 'returns task' do
+        expect(config.setup_lease_refresher).to eq(task)
+      end
+
+      it 'starts task' do
+        config.setup_lease_refresher
+        expect(task).to have_received(:execute)
+      end
+    end
+
+    context 'with previous task' do
+      let(:previous_task) { instance_double(Concurrent::TimerTask, shutdown: true) }
+
+      it 'instantiate task with correct params' do
+        config.setup_lease_refresher(previous_task)
+        expect(Concurrent::TimerTask).to have_received(:new).with(execution_interval: config.lease_refresh_interval)
+      end
+
+      it 'returns task' do
+        expect(config.setup_lease_refresher(previous_task)).to eq(task)
+      end
+
+      it 'starts task' do
+        config.setup_lease_refresher(previous_task)
+        expect(task).to have_received(:execute)
+      end
+
+      it 'shuts down old task' do
+        config.setup_lease_refresher(previous_task)
+        expect(previous_task).to have_received(:shutdown)
+      end
+    end
+  end
+end